+++ test/unit-tests/core/cfn-functions/cfn-json-string.test.ts
@@ -24,6 +24,26 @@ describe('when invoking cfn json string', () => {
         expect(element.key).toBe('{"att":"val","n":1}');
     });
 
+    test('JSON string gets minimized passed as 1st arg of array', async () => {
+        const element: any = {};
+        const json = `{
+            "key": "val"
+        }`;
+        CfnJsonString.resolve(null, {}, element, 'key', 'Fn::JsonString', [json])
+
+        expect(element.key).toBe('{"key":"val"}');
+    });
+
+    test('object gets converted to json passed as 1st arg of array', async () => {
+        const element: any = {};
+        const obj = {
+            att: 'val',
+            n: 1
+        };
+        CfnJsonString.resolve(null, {}, element, 'key', 'Fn::JsonString', [obj])
+
+        expect(element.key).toBe('{"att":"val","n":1}');
+    });
     test('JSON string can be pretty printed', async () => {
         const element: any = {};
         const json = `{