+++ CHANGELOG.md
@@ -3,7 +3,10 @@ All notable changes to aws organization formation will be documented in this fil
 
 **(Unreleased) version 0.9.12**
 - Allow failure tolerance to be set to 0 on validate-tasks command (allows CI/CD processes to fail on validation)
-- Added support for `Mappings`/`FindInMap` for task files.
+- Added support for `Mappings` / `FindInMap` for task files.
+- Added functions `!MD5` / `!ReadFile` that can be used in task files.
+- Added function `!JsonString` that can be used in task files.
+- Added support for `!Ref OrganizationRoot` (and other types) in task files.
 - Fixed bug on `org-formation init` where tags on the MasterAccount where not added to generated template.
 - Updating stacks that have state ROLLBACK_FAILED will be retried.